+++ pkg/apis/release/v1alpha1/release_types.go
@@ -17,9 +17,9 @@ const (
 type ReleaseState string
 
 var (
-	stateActive     ReleaseState = "active"     // nolint
-	stateDeprecated ReleaseState = "deprecated" // nolint
-	stateWIP        ReleaseState = "wip"        // nolint
+	StateActive     ReleaseState = "active"
+	StateDeprecated ReleaseState = "deprecated"
+	StateWIP        ReleaseState = "wip"
 )
 
 func (r ReleaseState) String() string {
@@ -128,6 +128,8 @@ type ReleaseSpecApp struct {
 type ReleaseStatus struct {
 	// Ready indicates if all components of the release have been deployed.
 	Ready bool `json:"ready"`
+	// InUse indicates whether a release is actually used by a cluster.
+	InUse bool `json:"inUse"`
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object