+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `InUse` to the `ReleaseStatus`.
+
 ## [2.1.0] - 2020-08-17
 
 ### Added